Python Project File On Bike Rental Prediction
Sharing Bike has come a new concept now a days where one doesn’t have to buy a bike to enjoy their daily ride. Currently It has introduced in so many urban countries for enhancement of mobility comfort. One can rent bikes on several basis like hourly, daily, monthly etc. It has a significant role to the rising issues related to the global warming, climate change, carbon emission and many more environmental anomalies. It is very important to make rental bikes available and accessible to the customers to the right time as it lessens the waiting time. In this project, we choose to analyse a dataset containing the rental bike count and other climate related variables of Seoul city, The capital of South Korea. First, we have focused on the data preprocessing and data transformation, then we have used “Linear Regression” to predict the rent bike count required at each hour very efficiently.